+++ packages/ui/src/components/index.ts
@@ -58,3 +58,5 @@ export type {
 
 export * from './InfoTable';
 export * from './Dropdown';
+export * from './Loading';
+export * from './Logo';