+++ CHANGELOG.md
@@ -2,12 +2,13 @@
 
 ### vNext
 
-- DAV-2958
-- DAV-2826
+- [DAV-3012] Wallet application screens layout and navigation
+- [DAV-3029] Primary color adjusted according to design
 
 ### v1.1.0
 
-- 
+- [DAV-2787] Transaction confirmation popups
+- [DAV-2826] Add blocked popup authorization request
 
 ### v1.0.0